Created a View class that the SlickApp calls < organisation
