+++ pom.xml
@@ -40,7 +40,7 @@ xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xs
                     </descriptorRefs>
                     <archive>
                         <manifest>
-                            <mainClass>com.sem.btrouble.view.SlickApp</mainClass>
+                            <mainClass>com.sem.btrouble.SlickApp</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
